- callback handling moved to seperate go file
