+++ drv.go
@@ -72,7 +72,6 @@ package godror
 
 #include "dpi.c"
 
-void TokenCallbackHandlerDebug(void* context, dpiAccessToken *token);
 */
 import "C"
 
@@ -84,7 +83,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"math"
 	"runtime"
 	"strconv"
@@ -777,41 +775,6 @@ func (d *drv) getPool(P commonAndPoolParams) (*connPool, error) {
 	return pool, nil
 }
 
-// AccessToken Callback information.
-type AccesTokenCB struct {
-	//pool     *connPool
-	callback func(*dsn.AccesToken)
-	ID       uint64
-}
-
-// Cannot pass *AccesTokenCB to C, so pass an uint64 that points to this map entry
-var (
-	accessTokenMu sync.Mutex
-	accessTokens  = make(map[uint64]*AccesTokenCB)
-	accessTokenID uint64
-)
-
-// tokenCallbackHandler is the callback for C code on token expiry.
-//
-//export TokenCallbackHandler
-func TokenCallbackHandler(ctx unsafe.Pointer, accessTokenC *C.dpiAccessToken) {
-	log.Printf("CB %p %+v", ctx, accessTokenC)
-	accessTokenMu.Lock()
-	acessTokenCB := accessTokens[*((*uint64)(ctx))]
-	accessTokenMu.Unlock()
-
-	// Call user function which provides the token and privateKey.
-	var at dsn.AccesToken
-	acessTokenCB.callback(&at)
-
-	token := at.Token
-	privateKey := at.PrivateKey
-	accessTokenC.token = C.CString(token)
-	accessTokenC.tokenLength = C.uint32_t(len(token))
-	accessTokenC.privateKey = C.CString(privateKey)
-	accessTokenC.privateKeyLength = C.uint32_t(len(privateKey))
-}
-
 // createPool creates an ODPI-C pool with the specified parameters.
 //
 // This is done while holding the mutex in order to ensure that
@@ -902,17 +865,7 @@ func (d *drv) createPool(P commonAndPoolParams) (*connPool, error) {
 	if P.TokenCB != nil {
 		//typedef int (*dpiAccessTokenCallback)(void *context,
 		//    dpiAccessToken *accessToken);
-		poolCreateParams.accessTokenCallback = C.dpiAccessTokenCallback(C.TokenCallbackHandlerDebug)
-		// cannot pass &token to C, so pass indirectly
-		aToken := AccesTokenCB{callback: P.TokenCB}
-		accessTokenMu.Lock()
-		accessTokenID++
-		aToken.ID = accessTokenID
-		accessTokens[aToken.ID] = &aToken
-		accessTokenMu.Unlock()
-		tokenID := (*C.uint64_t)(C.malloc(8))
-		*tokenID = C.uint64_t(accessTokenID)
-		poolCreateParams.accessTokenCallbackContext = unsafe.Pointer(tokenID)
+		RegisterTokenCallback(poolCreateParams, P.TokenCB)
 	}
 
 	// setup credentials
